feat: sign and more about authorization

+++ sns/server/controller/posts.ts
@@ -1,7 +1,12 @@
 import * as postRepository from '../model/posts';
 import { Request, Response } from 'express';
 
-type httpFunction = (req: Request, res: Response) => Promise<any>;
+type authRequest = Request & {
+	userId?: number;
+	token?: string;
+};
+
+type httpFunction = (req: authRequest, res: Response) => Promise<any>;
 
 export const getPosts: httpFunction = async (req, res) => {
 	const { userId } = req.query;
@@ -23,8 +28,12 @@ export const getPost: httpFunction = async (req, res) => {
 };
 
 export const createPost: httpFunction = async (req, res) => {
-	const { msg, userId } = req.body;
-	const post = await postRepository.create(msg, userId);
+	const { msg } = req.body;
+	if (typeof req.userId !== 'number' && !req.userId) {
+		return res.status(401).json({ message: 'Authentication Error' });
+	}
+
+	const post = await postRepository.create(msg, req.userId);
 
 	res.status(201).json(post);
 };
@@ -32,11 +41,18 @@ export const createPost: httpFunction = async (req, res) => {
 export const updatePost: httpFunction = async (req, res) => {
 	const { id } = req.params;
 	const { msg } = req.body;
-	const post = await postRepository.update(Number(id), msg);
+	const post = await postRepository.getById(Number(id));
+	if (!post) {
+		return res.sendStatus(404);
+	}
+	if (post.userId !== req.userId) {
+		return res.sendStatus(403);
+	}
+	const updated = await postRepository.update(Number(id), msg);
 
-	if (post) {
-		console.log(post);
-		res.status(200).json(post);
+	if (updated) {
+		console.log(updated);
+		res.status(200).json(updated);
 	} else {
 		res.status(404).json({ message: `post id(${id}) not found` });
 	}
@@ -44,6 +60,14 @@ export const updatePost: httpFunction = async (req, res) => {
 
 export const deletePost: httpFunction = async (req, res) => {
 	const { id } = req.params;
+	const post = await postRepository.getById(Number(id));
+	if (!post) {
+		return res.sendStatus(404);
+	}
+	if (post.userId !== req.userId) {
+		return res.sendStatus(403);
+	}
+
 	await postRepository.remove(Number(id));
 
 	res.sendStatus(204);
